requirements.txt
openai
anthropic
google-generativeai
click
Pygments
gnureadline;platform_system=="Linux" or platform_system=="Darwin"
requests
bs4
beautifulsoup4
tiktoken
reportlab
trafilatura

